+++ src/lib/index.ts
@@ -1,0 +1,5 @@
+export { default as camelCase } from './camelCase';
+export { default as createClassName } from './createClassName';
+export { default as isFileIncluded } from './isFileIncluded';
+export * from './getLocalIdent';
+export * from './patterns';