+++ test/compiler.js
@@ -1,14 +1,10 @@
 const svelte = require('svelte/compiler');
-const cssModules = require('../index.js');
+const cssModules = require('../dist/index.js');
 
 module.exports = async ({ source }, options) => {
-  const { code } = await svelte.preprocess(
-    source,
-    [
-      cssModules(options)
-    ],
-    { filename : 'src/App.svelte' }
-  );
+  const { code } = await svelte.preprocess(source, [cssModules(options)], {
+    filename: 'test/App.svelte',
+  });
 
   return code;
 };