+++ test/stylesFixtures/chaining-selector.test.js
@@ -1,4 +1,4 @@
-const compiler = require('./compiler.js');
+const compiler = require('../compiler.js');
 
 describe('combining multiple classes', () => {
   const style = '<style>span.red.large:hover { font-size: 20px; } \n.red { color: red; }</style>';
@@ -23,19 +23,18 @@ describe('combining multiple classes', () => {
 });
 
 describe('Classname is part of a selector', () => {
-  
   test('CSS Modules class targetting children', async () => {
     const source =
       '<style>\n' +
-        'div.red > sup { font-size: 12px; }\n' +
-        '.red { color: red; }\n' +
+      'div.red > sup { font-size: 12px; }\n' +
+      '.red { color: red; }\n' +
       '</style>\n' +
       '<div class="$style.red">Red<sup>*</sup></div>';
 
     const expectedOutput =
       '<style>\n' +
-        ':global(div.red-123) > sup { font-size: 12px; }\n' +
-        ':global(.red-123) { color: red; }\n' +
+      ':global(div.red-123) > sup { font-size: 12px; }\n' +
+      ':global(.red-123) { color: red; }\n' +
       '</style>\n' +
       '<div class="red-123">Red<sup>*</sup></div>';
 
@@ -54,17 +53,17 @@ describe('Classname is part of a selector', () => {
   test('CSS Modules class has a parent', async () => {
     const source =
       '<style>\n' +
-        'div .semibold .red { font-size: 20px; }\n' +
-        '.red { color: red; }\n' +
-        '.semibold { font-weight: 600; }\n' +
+      'div .semibold .red { font-size: 20px; }\n' +
+      '.red { color: red; }\n' +
+      '.semibold { font-weight: 600; }\n' +
       '</style>\n' +
       '<div><strong class="$style.semibold"><span class="$style.red">Red</span></strong></div>';
-  
+
     const expectedOutput =
       '<style>\n' +
-        'div :global(.semibold-123) :global(.red-123) { font-size: 20px; }\n' +
-        ':global(.red-123) { color: red; }\n' +
-        ':global(.semibold-123) { font-weight: 600; }\n' +
+      'div :global(.semibold-123) :global(.red-123) { font-size: 20px; }\n' +
+      ':global(.red-123) { color: red; }\n' +
+      ':global(.semibold-123) { font-weight: 600; }\n' +
       '</style>\n' +
       '<div><strong class="semibold-123"><span class="red-123">Red</span></strong></div>';
 
@@ -83,15 +82,15 @@ describe('Classname is part of a selector', () => {
   test('CSS Modules class has a global parent', async () => {
     const source =
       '<style>\n' +
-        ':global(div) .red { font-size: 20px; }\n' +
-        '.red { color: red; }\n' +
+      ':global(div) .red { font-size: 20px; }\n' +
+      '.red { color: red; }\n' +
       '</style>\n' +
       '<div><span class="$style.red">Red</span></div>';
-  
+
     const expectedOutput =
       '<style>\n' +
-        ':global(div) :global(.red-123) { font-size: 20px; }\n' +
-        ':global(.red-123) { color: red; }\n' +
+      ':global(div) :global(.red-123) { font-size: 20px; }\n' +
+      ':global(.red-123) { color: red; }\n' +
       '</style>\n' +
       '<div><span class="red-123">Red</span></div>';
 
@@ -110,19 +109,19 @@ describe('Classname is part of a selector', () => {
   test('CSS Modules class is used within a media query', async () => {
     const source =
       '<style>\n' +
-        '@media (min-width: 37.5em) {\n' +
-          '.red { color: red; }\n' +
-          'div.bold { font-weight: bold; }\n' +
-        '}\n' +
+      '@media (min-width: 37.5em) {\n' +
+      '.red { color: red; }\n' +
+      'div.bold { font-weight: bold; }\n' +
+      '}\n' +
       '</style>\n' +
       '<div class="$style.bold"><span class="$style.red">Red</span></div>';
-  
+
     const expectedOutput =
       '<style>\n' +
-        '@media (min-width: 37.5em) {\n' +
-          ':global(.red-123) { color: red; }\n' +
-          ':global(div.bold-123) { font-weight: bold; }\n' +
-        '}\n' +
+      '@media (min-width: 37.5em) {\n' +
+      ':global(.red-123) { color: red; }\n' +
+      ':global(div.bold-123) { font-weight: bold; }\n' +
+      '}\n' +
       '</style>\n' +
       '<div class="bold-123"><span class="red-123">Red</span></div>';